requirements.txt
rich~=13.7.1
numpy~=1.26.4
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
dlib~=19.24.4
face-recognition~=1.3.0

pandas~=2.2.2
requests>=2.32.0
seaborn>=0.11.0
tqdm>=4.64.0
gitpython>=3.1.30
matplotlib>=3.3
pillow>=10.3.0
psutil  # system resources
pyyaml>=5.3.1
scipy>=1.4.1
thop>=0.1.1  # FLO
ultralytics~=8.2.22
setuptools>=65.5.1

